Fix input line buffering
Previously was only passing lines to the parser
when the stream finishes. Unfortunately the parser
relies on the line timing to generate the test
duration, and so this was always showing as 0.

Now send lines to the parser when the reader is
blocked. This means we get more accurate timing
information on tests and reduces overall memory
consumption.

Obviously this only works when streaming adb
output directly rather than from a file.